feat: 更新文档

+++ src/core/core.js
@@ -7,7 +7,6 @@ const fse = require("fs-extra");
 
 // 默认配置项由外层调用传入
 let _option = {
-  projectList: "", //项目名列表
   suffixReg: "", //打包后匹配后缀
   gitProjectMap: "", //项目映射
   outputPutDir: "", //输入目录
@@ -15,9 +14,19 @@ let _option = {
   buildName: "", //打包文件名
   compressType: "", //压缩类型
   workSpacesPath: "", //工作空间目录
-  isConcurrentExecute: "", //是否并发执行
   defaultNodeVersion: "", //node默认版本
+  isConcurrentExecute: false, //是否并发执行
+  isClearWorkSpace: false, //是否需要清理工作空间(优先级高于单个项目配置的清理字段isClearCache)
+  isZipBundleFile: false, //是否打包后压缩打包文件并且删除文件夹
 };
+const timeCache = {
+  startTime: "",
+  endTime: "",
+};
+// 根据文件名获取执行环境地址
+function getCwdPath(fileName) {
+  return path.resolve(_option.workSpacesPath, fileName);
+}
 // 执行shell脚本
 async function execCommand(command, cwd) {
   return new Promise((resolve, reject) => {
@@ -25,11 +34,13 @@ async function execCommand(command, cwd) {
       reject("shell脚本不能为空");
     }
     let dirCommand = spawn(command, [], {
-      shell:true,
+      shell: true,
       cwd: cwd || path.resolve(__dirname),
     });
+    let logInfo = [];
     // 处理标准输出
     dirCommand.stdout.on("data", (data) => {
+      logInfo.push(data.toString());
       chalkUtil.logInfo(data.toString());
     });
     // 处理标准错误输出
@@ -38,12 +49,12 @@ async function execCommand(command, cwd) {
     });
     // 处理命令执行结果
     dirCommand.on("close", (data) => {
-      resolve("");
+      resolve(logInfo);
     });
   });
 }
 // 文件拷贝
-async function copyFile(fileName) {
+async function copyFile(project, fileName) {
   return new Promise(async (resolve, reject) => {
     try {
       const targetPath = path.resolve(
@@ -78,13 +89,14 @@ async function copyFile(fileName) {
           hasCopy = true;
           resolve();
         } else {
-          // 拷贝文件夹(不存在压缩文件的情况下)
-          if (i === matchedFiles.length - 1) {
+          // 拷贝文件夹(不存在压缩文件的情况下),需要拷贝文件的则拷贝文件
+          if (project.isBundleFile || i === matchedFiles.length - 1) {
             fse.copySync(originSourcePath, targetSourcePath);
             console.log(
               "target-outpath",
               path.resolve(targetSourcePath, "./" + file)
             );
+            hasCopy = true;
             resolve();
           }
         }
@@ -96,23 +108,25 @@ async function copyFile(fileName) {
   });
 }
 // 打包项目
-async function bundle(fileName) {
+async function bundle(project, fileName) {
   return new Promise(async (resolve, reject) => {
     try {
       // 当前环境执行目录
-      const cwd = path.resolve(_option.workSpacesPath, fileName);
+      const cwd = getCwdPath(fileName);
       // 当前项目执行node版本
       const currentProjectNodeVersion =
-        _option.gitProjectMap[fileName].nodeVersion ||
-        _option.defaultNodeVersion;
+        project.nodeVersion || _option.defaultNodeVersion;
       chalkUtil.logSuccess("当前环境执行目录");
-      console.log("cwd---------", cwd);
       await execCommand("cd", cwd);
-      // 当前所处分支
-      chalkUtil.logSuccess("current branch");
-      await execCommand("git branch --show-current", cwd);
-      chalkUtil.logSuccess("current tag");
-      await execCommand("git describe --tags --exact-match", cwd);
+      if (project.branch) {
+        // 当前所处分支
+        chalkUtil.logSuccess("current branch");
+        await execCommand("git branch --show-current", cwd);
+      } else if (project.tag) {
+        // 当前所处tag
+        chalkUtil.logSuccess("current tag");
+        await execCommand("git describe --tags --exact-match", cwd);
+      }
       //  切换node版本
       chalkUtil.logSuccess("start nvm check -node -version");
       await execCommand("nvm use" + " " + currentProjectNodeVersion, cwd);
@@ -134,7 +148,7 @@ async function bundle(fileName) {
       chalkUtil.logSuccess("end build....");
       // 拷贝打包文件
       chalkUtil.logSuccess("start copy file..");
-      await copyFile(fileName);
+      await copyFile(project, fileName);
       chalkUtil.logSuccess("end copy file..");
       resolve("");
     } catch (error) {
@@ -142,44 +156,76 @@ async function bundle(fileName) {
     }
   });
 }
+// 更新tag或者分支
+async function updateLocalBranchOrTag(project, fileName) {
+  return new Promise(async (resolve, reject) => {
+    // 当前环境执行目录
+    const cwd = getCwdPath(fileName);
+    try {
+      // 更新代码
+      if (project.branch) {
+        // 当前为分支的情况
+        chalkUtil.logSuccess(" start git pull");
+        const updateBranchShell = "git pull";
+        await execCommand(updateBranchShell, cwd);
+        chalkUtil.logSuccess(" end git pull");
+      } else if (project.tag) {
+        // 当前为tag的情况
+        chalkUtil.logSuccess(" start git reset tag");
+        const updateTagShell = "git pull origin " + project.tag;
+        await execCommand(updateTagShell, cwd);
+        chalkUtil.logSuccess(" end git reset tag");
+      }
+
+      resolve();
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
 // git clone项目
-async function cloneProject(fileName) {
+async function cloneProject(project, fileName) {
   return new Promise(async (resolve, reject) => {
     try {
+      const cwd = getCwdPath(fileName);
       // 克隆项目
       chalkUtil.logSuccess("start git  clone " + fileName);
-      const cloneShell =
-        "git clone " +
-        _option.gitProjectMap[fileName].gitUrl +
-        " --branch " +
-        _option.gitProjectMap[fileName].branchOrTag;
+      const cloneShell = "git clone " + project.gitUrl;
       await execCommand(cloneShell, _option.workSpacesPath);
       chalkUtil.logSuccess("end git  clone " + fileName);
+      if (project.branch) {
+        // 设置上游分支和本地的联系
+        // chalkUtil.logSuccess("start set-upstream" + fileName);
+        // const currentUpStreamBranchShell = ` git branch --set-upstream-to=origin/${project.branch}  ${project.branch}`;
+        // await execCommand(currentUpStreamBranchShell, cwd);
+        // 切换分支
+        chalkUtil.logSuccess("start git  checkout branch " + fileName);
+        const currentCheckoutBranchShell = `git checkout  ${project.branch}`;
+        await execCommand(currentCheckoutBranchShell, cwd);
+        await updateLocalBranchOrTag(project, fileName);
+      } else if (project.tag) {
+        // 切换tag
+        chalkUtil.logSuccess("start git  checkout tag " + fileName);
+        const currentCheckoutTagShell = `git checkout  ${project.tag}`;
+        await execCommand(currentCheckoutTagShell, cwd);
+        await updateLocalBranchOrTag(project, fileName);
+      }
       // 是否需要合并分支
-      const mergeBranch = _option.gitProjectMap[fileName].mergeBranch;
+      const mergeBranch = project.mergeBranch;
       if (mergeBranch) {
         chalkUtil.logSuccess("start merge branch " + mergeBranch);
         const mergeBranchShell = `git merge ${mergeBranch}`;
-        await execCommand(
-          mergeBranchShell,
-          path.resolve(_option.workSpacesPath, fileName)
-        );
+        await execCommand(mergeBranchShell, cwd);
       }
       // 是否需要合并tag
-      const mergeTag = _option.gitProjectMap[fileName].mergeTag;
+      const mergeTag = project.mergeTag;
       if (mergeTag) {
         chalkUtil.logSuccess("merge tag " + mergeTag);
         // 获取所有tag
-        await execCommand(
-          "git fetch origin --tags",
-          path.resolve(_option.workSpacesPath, fileName)
-        );
+        await execCommand("git fetch origin --tags", cwd);
         // 合并tag
         const mergeTagShell = `git merge ${mergeTag}`;
-        await execCommand(
-          mergeTagShell,
-          path.resolve(_option.workSpacesPath, fileName)
-        );
+        await execCommand(mergeTagShell, cwd);
       }
       resolve();
     } catch (error) {
@@ -187,6 +233,28 @@ async function cloneProject(fileName) {
     }
   });
 }
+// 本地已存在项目则直接使用，不用拉取了
+async function existProjectCheckoutBranchOrTag(project, fileName) {
+  return new Promise(async (resolve, reject) => {
+    // 当前环境执行目录
+    const cwd = getCwdPath(fileName);
+    try {
+      // 先fetch远程最新仓库更新
+      chalkUtil.logSuccess("start git  fetch " + fileName);
+      const fetchShell = "git fetch";
+      await execCommand(fetchShell, cwd);
+      // 切换分支或者tag
+      const cloneShell = "git checkout " + (project.branch || project.tag);
+      await execCommand(cloneShell, cwd);
+      // 更新代码
+      await updateLocalBranchOrTag(project, fileName);
+      resolve();
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
+
 // 压缩文件
 function zipAsset() {
   return new Promise((resolve, reject) => {
@@ -220,7 +288,7 @@ function zipAsset() {
 // 删除对应文件或者文件夹内容文件
 function removeDir(removePath) {
   return new Promise((resolve, reject) => {
-    fs.rmdir(removePath, { recursive: true, force: true }, (err) => {
+    fse.remove(removePath, (err) => {
       if (err) {
         reject(err);
       }
@@ -228,6 +296,7 @@ function removeDir(removePath) {
     });
   });
 }
+
 // 创建工作空间
 async function createWorkSpaces() {
   return new Promise((resolve, reject) => {
@@ -244,20 +313,22 @@ async function createWorkSpaces() {
     }
   });
 }
-// 打包完成后输入文件的操作
+// 打包完成后输出文件的操作
 async function afterBundleOutPutResolve() {
   return new Promise(async (resolve, reject) => {
     try {
-      // 最后压缩文件
-      chalkUtil.logSuccess("start zip file..");
-      await zipAsset();
-      chalkUtil.logSuccess("end zip file..");
-      // 删除原来打包生成文件夹只保留压缩包
-      chalkUtil.logSuccess("start del bundle file..");
-      await removeDir(
-        path.resolve(_option.outputPutDir, _option.outputFileName)
-      );
-      chalkUtil.logSuccess("end del bundle file..");
+      if (_option.isZipBundleFile) {
+        // 最后压缩文件
+        chalkUtil.logSuccess("start zip file..");
+        await zipAsset();
+        chalkUtil.logSuccess("end zip file..");
+        // 删除原来打包生成文件夹只保留压缩包
+        chalkUtil.logSuccess("start del bundle file..");
+        await removeDir(
+          path.resolve(_option.outputPutDir, _option.outputFileName)
+        );
+        chalkUtil.logSuccess("end del bundle file..");
+      }
       resolve();
     } catch (error) {
       reject(error);
@@ -268,9 +339,14 @@ async function afterBundleOutPutResolve() {
 async function clearWorkSpace() {
   return new Promise(async (resolve, reject) => {
     try {
-      chalkUtil.logSuccess("start clear workSpace...");
-      await removeDir(path.resolve(_option.workSpacesPath));
-      chalkUtil.logSuccess("end clear workSpace...");
+      if (_option.isClearWorkSpace) {
+        chalkUtil.logSuccess("start clear workSpace...");
+        const removePath = path.resolve(_option.workSpacesPath);
+        await removeDir(removePath);
+        chalkUtil.logSuccess("end clear workSpace...");
+      } else {
+        chalkUtil.logSuccess("workSpace has  cached...");
+      }
       resolve();
     } catch (error) {
       reject(error);
@@ -281,47 +357,168 @@ async function clearWorkSpace() {
 async function concurrentExecute() {
   return new Promise(async (resolve, reject) => {
     try {
-      const cloneProjectAsyncList = _option.projectList.map((fileName) => {
-        return () => {
-          return cloneProject(fileName);
-        };
-      });
-      const bundleAsyncList = _option.projectList.map((fileName) => {
-        return () => {
-          return bundle(fileName);
-        };
-      });
+      let cloneProjectAsyncList = [];
+      let bundleAsyncList = [];
+      for (let fileName in _option.gitProjectMap) {
+        const project = _option.gitProjectMap[fileName];
+        if (!fs.existsSync(getCwdPath(fileName))) {
+          // 不存在就克隆
+          cloneProjectAsyncList.push(() => {
+            return cloneProject(project, fileName);
+          });
+        } else {
+          // 存在直接更新
+          chalkUtil.logSuccess(
+            "local workSpace exist clone project " + fileName
+          );
+          // 切换分支或者tag
+          await existProjectCheckoutBranchOrTag(project, fileName);
+        }
+        // 追加到打包队列
+        bundleAsyncList.push(() => {
+          return bundle(project, fileName);
+        });
+      }
       // 批量克隆项目
-      await Promise.all(cloneProjectAsyncList.map((fn) => fn()));
+      if (cloneProjectAsyncList.length) {
+        chalkUtil.logSuccess("batch clone  project start-----------");
+        await Promise.all(cloneProjectAsyncList.map((fn) => fn()));
+        chalkUtil.logSuccess("batch clone  project end-----------");
+      }
       //  批量打包
+      chalkUtil.logSuccess("batch bundle project start-----------");
       await Promise.all(bundleAsyncList.map((fn) => fn()));
-      //后续操作
-      await Promise.all([clearWorkSpace(), afterBundleOutPutResolve()]);
-      chalkUtil.logSuccess("success");
+      chalkUtil.logSuccess("end bundle project end-----------");
       resolve();
     } catch (error) {
       reject(error);
     }
   });
 }
+// 删除打包文件
+async function removeBundleFile(project, fileName) {
+  return new Promise(async (resolve, reject) => {
+    try {
+      // 清除本地缓存文件（删除整个克隆的项目）
+      if (project.isClearCache) {
+        chalkUtil.logSuccess("start remove cache file " + fileName);
+        await removeDir(getCwdPath(fileName));
+        chalkUtil.logSuccess("end remove cache file " + fileName);
+      } else {
+        // 只删除build文件夹
+        chalkUtil.logSuccess("start remove bundle file " + fileName);
+        await removeDir(
+          path.resolve(
+            _option.workSpacesPath,
+            fileName + "/" + _option.buildName
+          )
+        );
+        chalkUtil.logSuccess("end remove bundle file " + fileName);
+      }
 
+      resolve();
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
+// 批量删除缓存文件
+async function batchRemoveBundleFile() {
+  return new Promise(async (resolve, reject) => {
+    try {
+      let batchPromiseList = [];
+      for (let fileName in _option.gitProjectMap) {
+        batchPromiseList.push(function () {
+          return removeBundleFile(_option.gitProjectMap[fileName], fileName);
+        });
+      }
+      let bachListLen = batchPromiseList.length;
+      let fulfilled = 0;
+      while (batchPromiseList.length && fulfilled < bachListLen) {
+        let bufferItem = batchPromiseList.shift();
+        bufferItem().finally(() => {
+          fulfilled++;
+          if (fulfilled === bachListLen) {
+            resolve();
+          }
+        });
+      }
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
+// 重置分支
+// function resetBranch(fileName) {
+//   return new Promise(async (resolve, reject) => {
+//     try {
+//       // 当前环境执行目录
+//       const cwd = getCwdPath(fileName);
+//       chalkUtil.logSuccess("start reset branch..");
+//       // 切换分支或者tag
+//       const cloneShell = "git checkout master";
+//       await execCommand(cloneShell, cwd);
+//       chalkUtil.logSuccess("end  reset branch....");
+//       resolve();
+//     } catch (error) {
+//       reject(error);
+//     }
+//   });
+// }
+// 输出成功的日志
+function outputBundleProjectLog() {
+  return new Promise((resolve) => {
+    console.log("bundle completed list-------------------");
+    for (let fileName in _option.gitProjectMap) {
+      chalkUtil.spinner.succeed(fileName);
+    }
+    console.log("bundle completed list-------------------");
+    resolve();
+  });
+}
+// 检测是否设置node-sass下载源
+function checkIsSetSassMirror() {
+  return new Promise(async (resolve) => {
+    const sassBinaryAddress = "https://npm.taobao.org/mirrors/node-sass";
+    const sassBinarySite = await execCommand(
+      "npm config get sass_binary_site",
+      path.resolve(__dirname)
+    );
+    if (
+      !sassBinarySite.find(
+        (logInfo) => logInfo.indexOf(sassBinaryAddress) !== -1
+      )
+    ) {
+      // 设置node-sass下载源，不然很容易出错
+      chalkUtil.logSuccess("npm  config set sass_binary_site");
+      await execCommand(
+        "npm  config set sass_binary_site " + sassBinaryAddress,
+        path.resolve(__dirname)
+      );
+      chalkUtil.logSuccess("yarn  config set sass_binary_site");
+      await execCommand(
+        "yarn  config set sass_binary_site " + sassBinaryAddress,
+        path.resolve(__dirname)
+      );
+    } else {
+      chalkUtil.logSuccess("yarn and npm  sass_binary_site has been set");
+    }
+    resolve();
+  });
+}
 // 启动入口
 async function start(option) {
   if (option) {
     _option = option;
   }
   try {
-    // 设置node-sass下载源，不然很容易出错
-    chalkUtil.logSuccess("npm  config set sass_binary_site");
-    await execCommand(
-      "npm  config set sass_binary_site https://npm.taobao.org/mirrors/node-sass",
-      path.resolve(__dirname)
-    );
-    chalkUtil.logSuccess("yarn  config set sass_binary_site");
-    await execCommand(
-      "yarn  config set sass_binary_site https://npm.taobao.org/mirrors/node-sass",
-      path.resolve(__dirname)
-    );
+    const projectLen = Object.keys(_option.gitProjectMap).length;
+    if (!projectLen) {
+      throw Error("请传入打包项目！");
+    }
+    timeCache.startTime = Date.now();
+    // 检测是否设置node-sass下载源
+    await checkIsSetSassMirror();
     // 检查工作空间文件夹是否存在不存在先创建
     await createWorkSpaces();
     // 并发打包
@@ -329,19 +526,38 @@ async function start(option) {
       await concurrentExecute();
     } else {
       // 串行打包
-      for (let i = 0; i < _option.projectList.length; i++) {
-        // 克隆项目
-        await cloneProject(_option.projectList[i]);
-        // 开始打包
-        await bundle(_option.projectList[i]);
-        // 打包完成后续处理
-        if (i === _option.projectList.length - 1) {
-          //后续操作
-          await Promise.all([clearWorkSpace(), afterBundleOutPutResolve()]);
+      for (let fileName in _option.gitProjectMap) {
+        const currentProjectItem = _option.gitProjectMap[fileName];
+        // 检查本地工作空间是否存在克隆的项目,存在则原基础上进行操作
+        if (fs.existsSync(getCwdPath(fileName))) {
+          chalkUtil.logSuccess(
+            "local workSpace exist clone project " + fileName
+          );
+          // 切换分支或者tag
+          await existProjectCheckoutBranchOrTag(currentProjectItem, fileName);
+        } else {
+          // 克隆项目
+          await cloneProject(currentProjectItem, fileName);
         }
+        // 开始打包
+        await bundle(currentProjectItem, fileName);
+        // 然后分支切换重置到master
+        // resetBranch(fileName);
       }
     }
+    // 清理工作空间、压缩代码
+    await Promise.all([clearWorkSpace(), afterBundleOutPutResolve()]);
+    // 批量删除build文件
+    if (!_option.isClearWorkSpace) {
+      await batchRemoveBundleFile();
+    }
+    // 输出打包完后的日志
+    await outputBundleProjectLog();
+    timeCache.endTime = Date.now();
     chalkUtil.spinner.succeed("success");
+    chalkUtil.spinner.succeed(
+      chalkUtil.generateDurationTime(timeCache.startTime, timeCache.endTime)
+    );
   } catch (error) {
     chalkUtil.logError(error);
   }
